Rename to `RedisConstants`
