+++ packages/cli/src/services/orchestration/helpers.ts
@@ -2,7 +2,7 @@ import { Container } from 'typedi';
 import { jsonParse } from 'n8n-workflow';
 import { Logger } from '@/Logger';
 import type { RedisServiceCommandObject } from '../redis/RedisServiceCommands';
-import { COMMAND_REDIS_CHANNEL } from '../redis/RedisServiceHelper';
+import { COMMAND_REDIS_CHANNEL } from '../redis/RedisConstants';
 import * as os from 'os';
 
 export interface RedisServiceCommandLastReceived {
